k12: refactoring
- Moves macros and lanes into the `lanes` module
- Extracts a `KangarooTwelve`-like type with a `Digest`-like interface

Doesn't actually impl any traits yet: will circle back on that after the
2018 edition upgrade.
